Add an IOSuite to clean tests code
